Moved mutations and querys to their own package and renamed Query to LinkQuerys
