+++ src/main/java/com/sudobangbang/graphql/mutation/LinkMutations.java
@@ -1,4 +1,4 @@
-package com.sudobangbang.graphql.resolver;
+package com.sudobangbang.graphql.mutation;
 
 import com.sudobangbang.graphql.endpoint.AuthContext;
 import com.sudobangbang.graphql.model.*;
